Fix imports for unit tests
There have been changes in the unit test tree and imports on neutron
side. Since networking-plumgrid repo's test files have imports from
neutron, as a result gate-networking-plumgrid-python27 if failing on
Jenkins because of import issues. This commit fixes them in test files.
In future we should also change directory strucuture to match that
of neutron.

Closes-Bug: #1465199
Change-Id: I810167b26a4cd50c047c789734c70f760f8db307