Fix imports for unit tests and update subnet APIs
There have been changes in the unit test tree and imports on neutron
side. Since networking-plumgrid repo's test files have imports from
neutron, as a result gate-networking-plumgrid-python27 if failing on
Jenkins because of import issues. This commit fixes them in test files.

Failures were not only caused by the imports but also because of some
missing functionality and checking in subnet APIs e.g.
- Safely extract gateway IP from a CIDR i.e. check if gateway in subnet
and also if IP address in on the subnet
- Don't extract gateway IP if no IPs left on subnet.

In future we should also change directory strucuture to match that
of neutron.

Closes-Bug: #1465199
Change-Id: I810167b26a4cd50c047c789734c70f760f8db307
Signed-off-by: Muawia Khan <muawia@plumgrid.com>