Fix for race condition in security groups creation
There was a race condition in the creation of default security group,
which was exposed when neutron was able to create a default security
group and backend controller fails which results in inconsitency between
neutron and backend controller.
This fix will allow both neutron and backend controller to rollback
in case of any failure i.e. either on neutron side or backend
controller to maintain consistency between both. It also removes the
strict checking for one default security group per tenant and will allow
neutron to handle this scenario

Change-Id: Ic0de2596d723ceba601d60371d644f320d5e37be
Closes-Bug: 1471287