Updating state of floating IP
Currently, in the plugin when a floatingip is associated its status is
ACTIVE and when it is disassociated from a VM the status is still
ACTIVE, this behavior also caused some tempest failures.
Fix monitors the state of floating IP on any all updates i.e.
association/disassociation and handles the status accordingly i.e.
ACTIVE/DOWN.

Closes-Bug: #1471115
Change-Id: I018449aa2c5197989d83f94b5a0f618b4c9035b1