Add DB for provider networks
Provider network extension has been part
of networking-plumgrid (previously PLUMgrid plugin)
for many releases, however, DB was not added. This is
an issue and DB needs to be there to have UTs pass
and extension to work.

The need for DB is needed as provider network extension
in Neutron does not implement a DB. Leaving all the plugins
to implement on their own.

Also, added unit tests cases for the extension as they
should be enabled now.

Closes-Bug: #1515492
Change-Id: If1b2645c2dfd7a9a8e0abbd25f88b21e4359aa19